Finised Factory Implementation, added Enum for easy API selection
