+++ COVID19Py/Concrete.py
@@ -1,0 +1,129 @@
+from interfaces import ICOVID19
+
+from requests import get
+from typing import Dict, List
+import json
+
+class ConcreteAPIDefault(ICOVID19):
+    url = "https://covid-tracker-us.herokuapp.com"
+
+    def __init__(self, data_source):
+        print("url", self.url)
+        result = get(self.url+"/v2/sources")
+        result.raise_for_status()
+        sources = result.json()["sources"]
+
+        if data_source in sources:
+            self.data_source = data_source
+        else:
+            raise Exception("data source not found")
+
+        self.previousData = None
+        self.latestData = None
+
+    def _request(self, endpoint, params=None):
+        print("datasource",self.data_source)
+        if (params is None):
+            params = {}
+        res = get(self.url+endpoint, {**params, "source": self.data_source})
+        res.raise_for_status()
+        return res.json()
+    def _update(self, timelines):
+        latest = self.getLatest()
+        locations = self.getLocations(timelines)
+        if self.latestData:
+            self.previousData = self.latestData
+        self.latestData = {
+            "latest": latest,
+            "locations": locations
+        }
+
+    def getAll(self, timelines=False):
+        self._update(timelines)
+        return self.latestData
+
+    def getLatestChanges(self):
+        changes = None
+        if self.previousData:
+            changes = {
+                "confirmed": self.latestData["latest"]["confirmed"] - self.latestData["latest"]["confirmed"],
+                "deaths": self.latestData["latest"]["deaths"] - self.latestData["latest"]["deaths"],
+                "recovered": self.latestData["latest"]["recovered"] - self.latestData["latest"]["recovered"],
+            }
+        else:
+            changes = {
+                "confirmed": 0,
+                "deaths": 0,
+                "recovered": 0,
+            }
+        return changes
+
+    def getLatest(self) -> List[Dict[str, int]]:
+        """
+        :return: The latest amount of total confirmed cases, deaths, and recoveries.
+        """
+        data = self._request("/v2/latest")
+        return data["latest"]
+
+    def getLocations(self, timelines=False, rank_by: str = None) -> List[Dict]:
+        """
+        Gets all locations affected by COVID-19, as well as latest case data.
+        :param timelines: Whether timeline information should be returned as well.
+        :param rank_by: Category to rank results by. ex: confirmed
+        :return: List of dictionaries representing all affected locations.
+        """
+        data = None
+        if timelines:
+            data = self._request("/v2/locations", {"timelines": str(timelines).lower()})
+        else:
+            data = self._request("/v2/locations")
+
+        data = data["locations"]
+        
+        ranking_criteria = ['confirmed', 'deaths', 'recovered']
+        if rank_by is not None:
+            if rank_by not in ranking_criteria:
+                raise ValueError("Invalid ranking criteria. Expected one of: %s" % ranking_criteria)
+
+            ranked = sorted(data, key=lambda i: i['latest'][rank_by], reverse=True)
+            data = ranked
+
+        return data
+
+    def getLocationByCountryCode(self, country_code, timelines=False) -> List[Dict]:
+        """
+        :param country_code: String denoting the ISO 3166-1 alpha-2 code (https://en.wikipedia.org/wiki/ISO_3166-1_alpha-2) of the country
+        :param timelines: Whether timeline information should be returned as well.
+        :return: A list of areas that correspond to the country_code. If the country_code is invalid, it returns an empty list.
+        """
+        data = None
+        if timelines:
+            data = self._request("/v2/locations", {"country_code": country_code, "timelines": str(timelines).lower()})
+        else:
+            data = self._request("/v2/locations", {"country_code": country_code})
+        return data["locations"]
+    
+    def getLocationByCountry(self, country, timelines=False) -> List[Dict]:
+        """
+        :param country: String denoting name of the country
+        :param timelines: Whether timeline information should be returned as well.
+        :return: A list of areas that correspond to the country name. If the country is invalid, it returns an empty list.
+        """
+        data = None
+        if timelines:
+            data = self._request("/v2/locations", {"country": country, "timelines": str(timelines).lower()})
+        else:
+            data = self._request("/v2/locations", {"country": country})
+        return data["locations"]
+
+    def getLocationById(self, country_id: int):
+        """
+        :param country_id: Country Id, an int
+        :return: A dictionary with case information for the specified location.
+        """
+        data = self._request("/v2/locations/" + str(country_id))
+        return data["location"]
+
+
+class ConcreteAPIMirror(ConcreteAPIDefault):
+    url = "https://cvtapi.nl"
