+++ COVID19Py/__init__.py
@@ -1,1 +1,1 @@
-from .covid19 import COVID19
+from .covid19 import COVID19,Source
